backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
websockets==14.2
python-dotenv==1.1.1
httpx>=0.27.0
